Revert "Wj test"
