--- log.txt
@@ -1,10 +1,0 @@
-File /media/bot/xtrastore/hinterland/hinterlander-quick-read/src/main.tsx
-
-
-Warnings :
-1 : the definition is not identified
-2 : the definition is not identified
-3 : the definition is not identified
-5 : the definition is not identified
-
-Your trace has 0 errors and 4 warnings.
